+++ CollaboratedHRM/StepDefinition/LoginStepDefinition.cs
@@ -3,7 +3,7 @@ using CollaboratedHRM.Setup;
 using System;
 using TechTalk.SpecFlow;
 
-namespace CollaboratedHRM
+namespace CollaboratedHRM.StepDefinition
 {
     [Binding]
     public class LoginStepDefinition
@@ -11,7 +11,7 @@ namespace CollaboratedHRM
         Base _base;
         LoginPage _loginPage;
         HomePage _homePage;
-        public LoginStepDefinition(Base basee,LoginPage login, HomePage homePage)
+        public LoginStepDefinition(Base basee, LoginPage login, HomePage homePage)
         {
             _base = basee;
             _loginPage = login;
@@ -39,7 +39,7 @@ namespace CollaboratedHRM
         [When(@"user clicks on Login Button")]
         public void WhenUserClicksOnLoginButton()
         {
-           _loginPage.clickSignIN();
+            _loginPage.clickSignIN();
         }
 
         [Then(@"Then user is logged in successfully")]
